Frontend folder set up with Chakra UI Setup

+++ 01_MERN-Crash-Course/backend/controllers/product.controller.js
@@ -50,10 +50,14 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
 
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({ message: "Invalid product ID" });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ message: "Product deleted successfully" });
   } catch (error) {
-    res.status(404).json({ message: "Product not found" });
+    res.status(500).json({ message: "Error in deleting product" });
   }
 };
